+++ packages/libdxfrw-converter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mlightcad/libdxfrw-converter
 
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/data-model@1.1.8
+
 ## 2.0.7
 
 ### Patch Changes